reuniokit-api-main/reuniokit-api-main/requirements.txt
aiobotocore==2.5.4
aiohttp==3.8.6
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
botocore==1.31.17
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
fsspec==2023.9.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
pluggy==1.3.0
psycopg2-binary==2.9.9
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
s3fs==2023.9.2
s3transfer==0.7.0
six==1.16.0
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.17
waitress==2.1.2
Werkzeug==3.0.0
wrapt==1.15.0
yarl==1.9.2
